% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundsheff.R
\docType{package}
\name{boundsheff}
\alias{boundsheff}
\title{boundsheff: A package of area boundaries for Sheffield as simple features}
\description{
The boundsheff package provides area boundaries for Sheffield as simple
features, with data from ONS or SCC AGOL.
}
\section{Author}{

Laurie Platt \email{laurie.platt@sheffield.gov.uk}
}

\section{See Also}{

\url{https://github.com/scc-pi/boundsheff}
}

