% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_asc_localities}
\alias{sf_asc_localities}
\title{Sheffield's 7 ASC (Adult Social Care) locality boundaries.}
\format{
A simple feature with seven variables: \code{objectid},
\code{ca_number}, \code{ca_name}, \code{asc_name}, \code{X}, \code{Y},
and \code{geom}.
}
\usage{
sf_asc_localities
}
\description{
Combination of four Wards in each and the same as the Sheffield LAC
(Local Area Committee) boundaries. From the SCC Portal.
}
\keyword{datasets}
