% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_lsoa}
\alias{sf_lsoa}
\title{Sheffield's 345 LSOA boundaries.}
\format{
A simple feature with five variables: \code{LSOA11CD},
\code{LSOA11NM}, \code{LONG}, \code{LAT}, and \code{geoms}.
}
\usage{
sf_lsoa
}
\description{
Generalised boundaries from ONS Open Geography Portal.
}
\keyword{datasets}
