% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_msoa}
\alias{sf_msoa}
\title{Sheffield's 70 MSOA boundaries.}
\format{
A simple feature with five variables: \code{MSOA11CD},
\code{MSOA11NM}, \code{LONG}, \code{LAT}, and \code{geoms}.
}
\usage{
sf_msoa
}
\description{
Generalised boundaries from ONS Open Geography Portal.
}
\keyword{datasets}
