% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_nhood}
\alias{sf_nhood}
\title{Sheffield's 100 neighbourhood boundaries.}
\format{
A simple feature with seven variables: \code{objectid},
\code{oa_code}, \code{oa_district}, \code{oa_ward},  \code{oa_nhood_code},
\code{nhood_name}, \code{X}, \code{Y}, and \code{geom}.
}
\usage{
sf_nhood
}
\description{
From the SCC Portal.
}
\keyword{datasets}
