% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_oa}
\alias{sf_oa}
\title{Sheffield's 1,817 Output Area boundaries.}
\format{
A simple feature with two variables: \code{OA11CD},
and \code{geoms}.
}
\usage{
sf_oa
}
\description{
Generalised boundaries from ONS Open Geography Portal.
}
\keyword{datasets}
