% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sf_ward}
\alias{sf_ward}
\title{Sheffield's 28 Ward boundaries.}
\format{
A simple feature with five variables: \code{WD20CD},
\code{WD20NM}, \code{LONG}, \code{LAT}, and \code{geoms}.
}
\usage{
sf_ward
}
\description{
Generalised boundaries from ONS Open Geography Portal.
}
\keyword{datasets}
